Update Rock Paper Scissors RNG
Rock Paper Scissors Project from Codecademy. This game uses RNG/Random Generated Numbers to output cases 0-2 and converts them to "Rock", "Paper", or "Scissors" based on their value and returns this as the variable result. Both the player choice and computer choice are randomly generated, with functionality for a tie using IF statements. The program will log the choices and the result to the console.